Merge Account Support
Adds support for merging accounts, as well as suggesting
that users try another login method in case they have a
duplicate account, so it can be merged.

As part of supporting merge accounts, this adds support for
multiple revenue cat ids per user. This also fixes a few indexes
on email fields which only matched on case-sensitive queries,
of which there should be none as emails are case-insensitive.
Part of this process involved bumping rqdb to 1.3.0 to use the
explain() function to view query plans

This includes a first basic pass replacing model_dump_json()
with __pydantic_serializer__.to_json() when bytes are clearly desired
anyway, to avoid two pointless linear passes (bytes -> str -> bytes).
This also applies to constructing all responses, but that change is left
for another commit since it's a bit less obvious that it avoids the
decode/encode