Allow special upgrade targets for journal_chat
The idea behind restricting the value is to ensure they can't
just reuse the same slug for success as the upgrade slug and thus
bypass the Oseh+ restriction. However, we do sometimes want to
handle upgrading differently (e.g., changing the back button). It's
sufficient to avoid confusion to just ensure its some flow containing
the special string, which is clearly intended for a premium class when
they don't have Oseh+